+++ src/AsyncWebSocket_RP2040W.cpp
@@ -9,7 +9,7 @@
   Built by Khoi Hoang https://github.com/khoih-prog/AsyncWebServer_RP2040W
   Licensed under GPLv3 license
  
-  Version: 1.3.0
+  Version: 1.3.1
   
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -22,6 +22,7 @@
   1.2.0   K Hoang      03/10/2022 Option to use cString instead of String to save Heap
   1.2.1   K Hoang      05/10/2022 Don't need memmove(), String no longer destroyed
   1.3.0   K Hoang      10/10/2022 Fix crash when using AsyncWebSockets server
+  1.3.1   K Hoang      10/10/2022 Improve robustness of AsyncWebSockets server
  *****************************************************************************************************************************/
 
 #include "Arduino.h"
@@ -784,8 +785,14 @@ void AsyncWebSocketClient::_queueMessage(AsyncWebSocketMessage *dataMessage)
 
   if (_messageQueue.length() >= WS_MAX_QUEUED_MESSAGES)
   {
-    AWS_LOGERROR("ERROR: Too many messages queued");
+    AWS_LOGERROR("ERROR: Large MsQ");
     delete dataMessage;
+    
+    // KH, fix _messageQueue overflowed by discarding all in the queue
+    _messageQueue.free();
+       
+    delay(20);
+    //////
   }
   else
   {